tool_server_lite/requirements.txt
# 核心依赖
litellm>=1.0.0          # 统一的LLM接口
pyyaml>=6.0             # YAML配置文件解析
tiktoken>=0.5.0
virtualenv>=20.0.0      # 虚拟环境（兼容 Anaconda）

fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
chardet>=5.2.0
pdfplumber>=0.10.0
python-docx>=1.1.0
python-pptx>=0.6.21
crawl4ai>=0.3.0
ddgs>=1.0.0
pyyaml>=6.0.0
arxiv>=2.0.0
prompt_toolkit>=3.0.0
rich>=13.0.0
psutil>=5.9.0 

playwright>=1.40.0
